Remove compile warnings
Mostly by changing to explicit casts.  But check these: unsigned ;ong
to double IS a narrowing cast.  Also one warning remains:

src/FMC_FE3.cpp:654:61: warning: format ‘%ld’ expects argument of type
    ‘long int’, but argument 3 has type ‘long int*’ [-Wformat=]
    fprintf(fout, "\tFuelKey: %ld\n", MxDesc.FuelKey[i][j]);

    long  FuelKey   [eC_Sta] [NUM_FUEL_SIZES] [256];

So this is a 3d array but the code in FMC_FE3 only refers to the first 2
dimensions, so yes, this is a pointer, not a long.